simplify DocumentOptions
